pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lc_conductor"
version = "0.1.0"
description = "LC-Conductor: Core components for integrating Agentic frameworks with web UI components with optimizations for Livermore Computing."
authors = []
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "charge[all]>=0.1.0",
    "mcp>=1.10.0",
    "jsonschema>=4.17.3",
    "loguru>=0.7.0",
    "pydantic>=2.10.7",
    "click",
    "pre-commit",
    "black",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["lc_conductor*"]

[project.optional-dependencies]
autogen = ["autogen-agentchat>=0.7.0", "autogen-core>=0.7.0", "autogen-ext>=0.7.0", "openai>=2.0.0", "tiktoken>=0.10.0", "autogen-ext[openai]"]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "responses>=0.22.0",
    "httpx>=0.24.0",
    "requests-mock>=1.11.0",
    "coverage[toml]>=7.0.0"
]

all = ["lc-conductor[autogen,test]"]
